requirements.txt
flask
discord
pyyaml
